allow wrapper elements around array results
Generally, Canvas returns lists of results as a bare JSON array.  For a few API endpoints, though, the lists are wrapped in a wrapper element (for some reason - feels like an API inconsistency).  Allow people to provide that wrapper element's name, and seamlessly unwrap them into an array for them.